Fix up some stuff on CUDA
Shut up a useless NVCC compiler warning about 'useless comparison
against 0' when it sees a constant n in the reducer class template.

CUDA dotprod didn't work with num workgroups == 256, so fix it up
to base its number of workgroups on the input size. To get performance
roughly equzl to what it was before, increase the number of values
handled by each thread.